Mark started phone application as keep alive
Somehow WebAppManager does not forward the keep alive status of the parent application to
possible child applications.

We want to have the phone application (the UI part) all-time available in the background
to be sure we can start it as fast as possible once we have to because of an incoming call
for example.

Open-WebOS-DCO-1.0-Signed-Off-By: Simon Busch <morphis@gravedo.de>