Fix incoming status when call ends
When the call ends, the ofono object holding the isIncoming information can be disconnected from DBus before we try to read the status one last time.
As isIncoming is the only property that is not stored locally, it results in a potentially wrong isIncoming value.
The fix here is to memorize the isIncoming status on the JS stack as soon as possible.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>